Lots going on, due to malfunction
Main purpose was making the form submit working. Because it was not working, I first read the code against the example repo. Turns out the example repo is not the same as the repo in the video, so that caused noise.
Also, because I am using Rails 5.1, so, without jQuery, I had to add it with yarn. That caused an csrf_token error, which is normally taken care of by jquery_ujs. Now I send it with the form.